% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\encoding{UTF-8}
\name{av}
\alias{av}
\title{Flatten \code{...} Into an Atomic Vector of the Consituent Atomic Elements of \code{...}}
\usage{
av(...)
}
\arguments{
\item{...}{Objects to be flattened to an attomic vector.}
}
\value{
An atomic vector (or \code{NULL})
}
\description{
Flatten \code{...} Into an Atomic Vector of the Consituent Atomic Elements of \code{...}
}
\seealso{
Other utils: 
\code{\link{check_valid}()},
\code{\link{combo2props}()},
\code{\link{meets_errs}()},
\code{\link{spec2_combos}()},
\code{\link{spec2props}()}
}
\concept{utils}
