% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masterFunction.R
\name{BRCAPRO}
\alias{BRCAPRO}
\title{BRCAPRO}
\usage{
BRCAPRO(...)
}
\arguments{
\item{...}{Additional arguments to be passed into \code{PPP}.}
}
\description{
Runs PPP with \code{genes = c("BRCA1", "BRCA2")} and
\code{cancers = c("Breast", "Ovarian", "Contralateral")}, as in
\code{PPP:::MODELPARAMS$BRCAPRO}.
}
\examples{
data(test_fam_1)
BRCAPRO(test_fam_1, parallel = FALSE)
}
\seealso{
\code{\link{PPP}}

Other predefined: 
\code{\link{BRCAPRO5}()},
\code{\link{BRCAPRO6}()},
\code{\link{MMRPRO}()},
\code{\link{PanelPRO11}()},
\code{\link{PanelPRO22}()}
}
\concept{predefined}
