% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masterFunction.R
\name{BRCAPRO5}
\alias{BRCAPRO5}
\title{BRCAPRO-5}
\usage{
BRCAPRO5(...)
}
\arguments{
\item{...}{Additional arguments to be passed into \code{PPP}.}
}
\description{
Runs PPP with \code{genes = c("BRCA1", "BRCA2", "ATM", "CHEK2", "PALB2")} and
\code{cancers = c("Breast", "Ovarian", "Contralateral")}, as in
\code{PPP:::MODELPARAMS$BRCAPRO5}.
}
\examples{
data(test_fam_1)
BRCAPRO5(test_fam_1, parallel = FALSE)
}
\seealso{
\code{\link{PPP}}

Other predefined: 
\code{\link{BRCAPRO6}()},
\code{\link{BRCAPRO}()},
\code{\link{MMRPRO}()},
\code{\link{PanelPRO11}()},
\code{\link{PanelPRO22}()}
}
\concept{predefined}
