% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masterFunction.R
\name{BRCAPRO6}
\alias{BRCAPRO6}
\title{BRCAPRO-6}
\usage{
BRCAPRO6(...)
}
\arguments{
\item{...}{Additional arguments to be passed into \code{PPP}.}
}
\description{
Runs PPP with
\code{genes = c("BRCA1", "BRCA2", "MLH1", "MSH2", "MSH6", "CDKN2A")}
and \code{cancers = c("Breast", "Ovarian", "Contralateral")}, as in
\code{PPP:::MODELPARAMS$BRCAPRO6}.
}
\examples{
data(test_fam_1)
BRCAPRO6(test_fam_1, parallel = FALSE)
}
\seealso{
\code{\link{PPP}}

Other predefined: 
\code{\link{BRCAPRO5}()},
\code{\link{BRCAPRO}()},
\code{\link{MMRPRO}()},
\code{\link{PanelPRO11}()},
\code{\link{PanelPRO22}()}
}
\concept{predefined}
