% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppp.R
\name{decorate_pb}
\alias{decorate_pb}
\title{Decorate a function to update a progressbar}
\usage{
decorate_pb(f, pb = .pb)
}
\arguments{
\item{f}{a function}

\item{pb}{the name of the progressbar}
}
\value{
a decorated variant of the original function
}
\description{
Decorate a function to update a progressbar
}
