% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_group_for_polys.R
\name{find_groups_in_image}
\alias{find_groups_in_image}
\title{find_groups_in_image}
\usage{
find_groups_in_image(x, image_id, dist_buffer = 12)
}
\arguments{
\item{x}{All image in a sf object}

\item{image_id}{image_id on which to find groups}

\item{dist_buffer}{Numeric. Distance of the buffering area. The smaller the better.}
}
\description{
find_groups_in_image
}
