% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_group_for_polys.R
\name{find_top_groups}
\alias{find_top_groups}
\title{Find best group for each image_pol_id independently}
\usage{
find_top_groups(x, all_ids)
}
\arguments{
\item{x}{dataframe as issued from \code{\link{group_pixels_count}}}

\item{all_ids}{vector of all image_pol_id}
}
\description{
Find best group for each image_pol_id independently
}
\details{
If an image_pol_id has no group left, it is included in a group alone
}
