% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppp_package.R
\name{findppp}
\alias{findppp}
\title{PPP extractor function}
\arguments{
\item{country}{Provide a country code found in OECD's R package}

\item{year}{Provide the year that you wish to extract data for}
}
\description{
Check a country's PPP in a given year.
}
\details{
This function helps you solve trawling through Google for crucial information on PPP.

The function asks the user to fill in the country followed by the year of desired observation
should this not be an error it will return a tibble with information relevant to the query.
}
\examples{
To retrieve PPP data for a particular year, just specify the country then year
econdata("AUS","2001")
}
\keyword{PPP,}
\keyword{data,}
\keyword{economic}
\keyword{year}
