% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_group_for_polys.R
\name{group_pixels_count}
\alias{group_pixels_count}
\title{Count number of pixels by groups of image_pol_id from stack}
\usage{
group_pixels_count(x)
}
\arguments{
\item{x}{stack as issued from \code{\link{voronoi_stacker}}}
}
\description{
Count number of pixels by groups of image_pol_id from stack
}
