% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masterFunction.R
\name{MMRPRO}
\alias{MMRPRO}
\title{MMRPRO}
\usage{
MMRPRO(...)
}
\arguments{
\item{...}{Additional arguments to be passed into \code{PPP}.}
}
\description{
Runs PPP with \code{genes = c("MLH1", "MSH2", "MSH6")} and
\code{cancers = c("Colorectal", "Endometrial")}, as in
\code{PPP:::MODELPARAMS$MMRPRO}.
}
\examples{
data(test_fam_3)
MMRPRO(test_fam_3, parallel = FALSE)
}
\seealso{
\code{\link{PPP}}

Other predefined: 
\code{\link{BRCAPRO5}()},
\code{\link{BRCAPRO6}()},
\code{\link{BRCAPRO}()},
\code{\link{PanelPRO11}()},
\code{\link{PanelPRO22}()}
}
\concept{predefined}
