% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{naics_clean}
\alias{naics_clean}
\title{Clean NAICS dataframe for joining with PPP data}
\usage{
naics_clean(naics_df)
}
\arguments{
\item{naics_df}{The tibble returned by \code{naics_read()}}
}
\value{
A tidied tibble of NAICS codes by year release
}
\description{
Clean NAICS dataframe for joining with PPP data
}
