% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{naics_join}
\alias{naics_join}
\title{Perform hierarchical join of PPP loan data with NAICS legends}
\usage{
naics_join(ppp_df, naics_df)
}
\arguments{
\item{ppp_df}{The tibble of PPP loan data}

\item{naics_df}{The tibble resulting from \code{naics_clean()}}
}
\value{
A tibble of tidy PPP loan data with NAICS industry labels
}
\description{
This function will attempt to match PPP loan data with NAICS codes
hierarchically. I.e., it will first attempt to match with the latest NAICS
legend (2017), then the next most recent (2012), and so on. NAICS Codes
versions checked are those available for download on the US Census website:
2017, 2012, 2007, and 2002.
}
