% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppp-package.R
\docType{package}
\name{ppp-package}
\alias{ppp}
\alias{ppp-package}
\title{ppp: Example Plumber API Package}
\description{
This package accompanies a talk given at RStudio::conf 2020 about
    practical Plumber patterns (ppp).
}
\keyword{internal}
