% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppp_assemble.R
\name{ppp_assemble}
\alias{ppp_assemble}
\title{Assemble a tidied PPP data set for analysis}
\usage{
ppp_assemble(version = NULL)
}
\arguments{
\item{version}{Either 1 for most recent version or 2 for oldest data. If not provided will ask}
}
\value{
A tidy tibble of PPP loan data (~5.2M rows)
}
\description{
This function will download, clean and return a large dataframe in memory of
PPP data.
}
