% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{ppp_clean}
\alias{ppp_clean}
\title{Applies recommended cleaning, see docs for details.}
\usage{
ppp_clean(df)
}
\arguments{
\item{df}{The PPP dataframe}
}
\value{
A cleaned \code{tibble} of PPP data.
}
\description{
Applies recommended cleaning, see docs for details.
}
