% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{ppp_collect}
\alias{ppp_collect}
\title{Collect PPP data from DataKind Google Drive}
\usage{
ppp_collect(version = NULL)
}
\arguments{
\item{version}{Either 1 for most recent version or 2 for oldest data. If not provided will ask}
}
\description{
Collect PPP data from DataKind Google Drive
}
