% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppp_delete_raw_data.R
\name{ppp_delete_raw_data}
\alias{ppp_delete_raw_data}
\title{Remove all raw PPP data files}
\usage{
ppp_delete_raw_data()
}
\description{
Deletes all files created by \code{ppp_collect()} and  \code{naics_collect()}
}
