% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{ppp_read}
\alias{ppp_read}
\title{Read PPP loan data .CSV files into R}
\usage{
ppp_read(version = NULL)
}
\arguments{
\item{version}{Either 1 for most recent version or 2 for oldest data. If not provided will ask}
}
\value{
A \code{tibble} of PPP loan data, ready for cleaning or analysis.
}
\description{
Read PPP loan data .CSV files into R
}
