% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict_sale_price}
\alias{predict_sale_price}
\title{Predict sale price of a property}
\usage{
predict_sale_price(new_data, all = TRUE, log = FALSE)
}
\arguments{
\item{new_data}{A \code{data.frame} or equivalent containing features necessary
for predicting sale price}

\item{all}{Logical. Whether to return just predicted values or the
transoformed \code{new_data} input along with predicted values. Defalt is
TRUE.}

\item{log}{Logical. The model returns log10 sale price values. If this
parameter is FALSE (the default) then the values will be transformed back
before being returned.}
}
\value{
If \code{all} is \code{FALSE}, a vector containing the predicted
  values; if \code{all} is \code{TRUE}, a tibble containing the predicted
  values along with the transformed columns of \code{new_data}
}
\description{
Predict sale price of a property
}
\examples{
\dontrun{
predict_sale_price(new_data = jsonlite::read_json(system.file("plumber",
                                                              "model",
                                                              "test-data.json",
                                                              package = "ppp"),
                                                  simplifyVector = TRUE))
}
}
