% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{run_predict_api}
\alias{run_predict_api}
\title{Run the housing prediction API}
\usage{
run_predict_api(...)
}
\arguments{
\item{...}{Options passed to \code{plumber::plumb()$run()}}
}
\value{
A running Plumber API
}
\description{
Run the housing prediction API
}
\examples{
\dontrun{
run_predict_api()
run_predict_api(swagger = TRUE, port = 8000)
}
}
