% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppp.R
\name{spawn_pb}
\alias{spawn_pb}
\title{Spawn a progressbar}
\usage{
spawn_pb(x, .name = .pb, .times = 1)
}
\arguments{
\item{x}{a data.frame or tibble}

\item{.name}{the name of the progressbar}

\item{.times}{multiply number of steps}
}
\value{
returns x unchanged
}
\description{
Spawn a progressbar
}
