% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec2combos.R
\encoding{UTF-8}
\name{spec2_combos}
\alias{spec2_combos}
\title{Convert a Property Spec to a Vector of Property Combos}
\usage{
spec2_combos(x)
}
\arguments{
\item{x}{A character vector or scalar.}
}
\value{
A character vector or \code{NULL} (returns \code{NULL} when \code{x} is not of mode \code{'character'} or is of length zero).
}
\description{
Splits a property spec along pipes to place each single property or property combo in a separate element (property combos are multiple properties separated by underscores).
}
\seealso{
Other utils: 
\code{\link{av}()},
\code{\link{check_valid}()},
\code{\link{combo2props}()},
\code{\link{meets_errs}()},
\code{\link{spec2props}()}
}
\concept{utils}
