% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_group_for_polys.R
\name{test_groups_kept}
\alias{test_groups_kept}
\title{Test if some image_pol_id are found in multiple groups}
\usage{
test_groups_kept(x)
}
\arguments{
\item{x}{dataframe as issued from \code{\link{find_top_groups}}}
}
\description{
Test if some image_pol_id are found in multiple groups
}
