% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_carto.R
\name{to_carto_grouped_sf}
\alias{to_carto_grouped_sf}
\title{Transform as grouped sf}
\usage{
to_carto_grouped_sf(x)
}
\arguments{
\item{x}{dataset with geometry column}
}
\description{
Transform as grouped sf
}
