% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_group_for_polys.R
\name{voronoi_rasterize}
\alias{voronoi_rasterize}
\title{Create raster Stack of voronoi and buffer around sf}
\usage{
voronoi_rasterize(x, dist_buffer = 12, r_image, keep_intermediate = FALSE)
}
\arguments{
\item{x}{sf object for a unique user}

\item{dist_buffer}{distance in pixel to crop voronoi around sf object}

\item{r_image}{raster object of the complete image to be the basis for rasterizing sf object}

\item{keep_intermediate}{Logical. If TRUE, return a list intermediate spatial objects}
}
\description{
Create raster Stack of voronoi and buffer around sf
}
